+++ packages/conferer/CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [PVP](https://pvp.haskell.org/).
 
 ## [Unreleased]
 
-Nothing
+* Add `mkConfig'` which allows creating a config by passing a list of defaults and
+a list of source creators.
+* Add `addSources`, which allows to add several sources to a config.
+* Define the type `Defaults`, which is a list of associations from `Key` to
+`Dynamic`.
 
 ## [v1.0.0.0] - 2020-12-29